webapps/1/requirements.txt
tensorflow
numpy
streamlit
pillow
matplotlib